+++ src/main/java/subway/application/StationService.java
@@ -1,7 +1,11 @@
-package subway;
+package subway.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.domain.Station;
+import subway.infrastructure.StationRepository;
+import subway.presentation.StationRequest;
+import subway.presentation.StationResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
