+++ src/main/java/subway/presentation/StationResponse.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.presentation;
 
 public class StationResponse {
     private Long id;
